Add --delimiter option for 'listpages' and 'getpagesummary'
Both of these actions used to have ", " as a hard-coded field
delimiter. This made it difficult to process the output if, eg.,
a page name had ", " in it. The new --delimiter option makes it
possible to use an arbitrary string as field delimiter, instead.

The delimiter defaults to ", " to avoid breaking things for
existing users.